Add Docker support
This commit adds a Dockerfile that will allow users to execute bgpq4
with Docker. This is particularly helpful in environments in which the
tools to compile the program are either unavailable or unwanted.

The application can be executed with something like:

```
docker build -t bswinnerton/bgpq4:latest .
docker run bswinnerton/bgpq4:latest AS36459
```